feat: add day 3
We're back for day 3! When I saw this one, I knew it had to be done with
regex, so my first goal was doing that in lisp. Unfortunately, common
lisp doesn't come with any way to do that so I had to add my first
external dependency - cl-ppcre via quicklisp. It's fine, but doesn't
have the niceties I'm used to (e.g. getting all matches and their groups
at once) leaving me splitting the regex into 2: a match part and a group
part.

That was all done fairly swiftly, though, leaving me with
not-quite-enough time for part 2. After some humming and harring I
settled on the solution you see before you - something presumably
similar to what I might write in python: a loop where I continuously
mutate a string and flip back and forth between an enabled and a
disabled state. Once I had realized this was possible, it was fairly
easy to implement (although not without prog vs progn vs block foibles),
however I spent an unreasonable amount of time wondering if this was
sensibly possible with pure regex (I presume it's solvable with negative
lookaheads) and also trying to think of a recursive solution that isn't
horrible (this algorithm recursively would -in my opinion- be a bit
horrible)